Update explainer.md
Adds a use case (low-latency streaming with a seekable buffer).
Moves decoded buffering support out-of-scope and to addendum.
Refines the title to be "Media Source Extensions for WebCodecs" (yes, there's a play on words in there.)
Notes that prototype will attempt to resolve a feasible way to support out-of-order WebCodecs encoded chunks (that have no decode timestamp.)
Removes out-of-scoped decoded frame support's example (it was remarkably similar to the encoded chunk support's example, too.)
Adds an open question about "Dictionary of Dictionaries" SourceBufferConfig ergonomics vs sniffing. Notes that TAG will be consulted about this.